social_media_api/requirements.txt
Django==3.2.9
djangorestframework==3.12.4
pillow==10.4.0